service/consumer/requirements.txt
pika==1.3.2
psycopg2-binary==2.9.9
pydantic==2.8.2
python-dotenv==1.0.1